Fix: bugs in TrashedRegisterFinder and DistributedConversionRule.
Reko was generating invalid CONVERT() expression.

+++ subjects/Elf/Sparc/rtems/sparc-rtems-unprotoize.reko/sparc-rtems-unprotoize_text.c
@@ -680,7 +680,7 @@ l00011DA0:
 		while (true)
 		{
 			Eq_n o0_n = strlen(l0_n);
-			if ((int8) (i0_n->t0000.u0 - *l0_n) == 0x00)
+			if ((int32) (i0_n->t0000.u0 - *l0_n) == 0x00)
 			{
 				if (strncmp(i0_n, l0_n, o0_n) == 0x00)
 				{
